Handle delete, get, and update itemByID requests.
